[lib/keyserver] make localID optional in the sendReactionMessageRequestInputValidator
Summary:
make `localID` optional in the `sendReactionMessageRequestInputValidator`. We need to do this because old `native` clients ie. build 172 does not have a localID field and by previously making `localID` requried it would cause the `sendReactionMessageRequestInputValidator` to fail

Depends on D6244

Test Plan:
Tested both older and newer `native` clients.

To test newer clients, I sent a reaction message and everything worked as expected, I checked the DB and the reaction message had the `creation` column filled.

To test older clients, I checked out the commit with the release tag using `git checkout v1.0.172`. This is to test an older `native` client. Then I copied and pasted the necessary code in the `message-responders` file to demonstrate an updated keyserver, and when I tried sending a reaction message, everything worked as expected and when I checked the DB, the reaction message did not have the `creation` column filled which makes sense since no `localID` was passed

{F326293}

Reviewers: atul, ashoat, tomek

Reviewed By: ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6245