[services] Tunnelbroker - Add skip authentication config parameter
Summary:
This diff introduces adding the `sessions.skip_authentication` config flag. It will be used in D6253 to skip the authentication mechanism.
As we don't have a function to get boolean flags from the config file (we have string parameters), the internal `isParameterSet` function was created to get the boolean value of the flag.

Linear task: [[ https://linear.app/comm/issue/ENG-2642/disabling-authentication-mechanism-by-the-config-flag-in-tunnelbroker | ENG-2642 ]]

Test Plan: Adding the `skip_authentication` flag to the `[sessions]` section of the config file results in a `true` return from the `isParameterSet('sessions.skip_authentication')` function call.

Reviewers: jon, bartek

Reviewed By: jon

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6251