[native] Add 'internal-ip' dependency
Summary:
Added `internal-ip` to `devDependencies` in native `package.json`.
I used the same exact version as the `@expo/cli` uses, so the `yarn.lock` remains unchanged.

The latest version of this library (`7.0.0`) is ESM-only, this is why the old `4.3.0` is used.

Test Plan: `yarn cleaninstall` doesn't do any changes to `yarn.lock`

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: tomek, atul

Differential Revision: https://phab.comm.dev/D6266