Send broadcast with new device token from CommNotificationsHandler to CommAndroidNotificationsEventEmitter broadcast receiver and forward it fo JavaScript where it is handled as previously
Summary:
This differential overrides `onNewToken` method of `FirebaseMessagineService` to send broadcast to CommAndroidNotificationsEventEmitter with new token. CommAndroidNotificationsEventEmitter receives broadcast with token and send event to JavaScript. On the Js side we register to listen to
this type of event and react to it in the same way we used to.

Test Plan: Build the app. Notifications should be delivered, since the device token should be correctly received.

Reviewers: tomek, bartek

Reviewed By: tomek, bartek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6272