Implement parsing notification received when application is foregrounded into object that can be directly passed to JavaScript. Implement relevant broadcast receiver as well
Summary:
This differential implements function that parses data received in CommNotificationsHandler when application is in the foreground into object that directly maps to what we expect to receive on the JavaScript side to display foreground notification. Relevant broadcast receiver is implemented
as well.

Test Plan: Build the app. Temporarily make parsing method as public static and call it in CommNotificationsHandler on `RemoteMessage` instance. App must not crash.

Reviewers: tomek, bartek

Reviewed By: tomek, bartek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6273