[web] Add tab to payload for navigating to a chat
Summary:
The function returned by `useOnClickThread` is used to navigate to a chat. Before, when called from a different tab that Chat, it would change the chat displayed in Chat tab, but not chagne the
tab to the Chat tab. ie if we callled it from Apps, the user would still see the Apps tab. Now calling this function will also change the tab to the Chat tab.

Test Plan:
Run web app, check that in Chat tab pressing a chat item in the chat thread list still navigates correctly to the corresponding chat. Check that calling this function from different tabs
changes the tab to the Chat tab, and navigates to the corresponding chat.

Reviewers:
przemek, michal, bartek

Subscribers: