[cargo] Bump `tokio` to `1.23.1`
Summary: To address the issues Dependabot is complaining about. Assuming `tokio` is important enough to the Rust ecosystem that they conform to semantic versioning stuff and this change is benign. Can roll back if there are any issues. Mostly just wanted to do this because I've been getting a bunch of emails about this.

Test Plan: NA, trusting `tokio` and CI

Reviewers: bartek, jon, tomek, varun, max

Reviewed By: jon

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D6290