Modify notification displayed from Java so that it opens the app when clicked
Summary: Previous differential in this stack disable the functionality that we open the app and navigate to relevant thread when notification is tapped. This differential brings back the first half of this functionality - tapping notification opens the app.

Test Plan: Build the app. Kill it. Send notification. Tap it. Ensure app opens (but not to the thread yet).

Reviewers: tomek, bartek

Reviewed By: tomek, bartek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6295