[native] Fix 'expo' command in nix env
Summary:
Analogous to D5508. I use the `expo run:android` instead of `react-native run-android` ([[ https://blog.expo.dev/introducing-expo-run-commands-835ae8da4813 | see why ]]) and it fails in Nix workflow. The solution is to do the same as in D5508.

I also added the `run-ios` and `run-android` scripts, because the `COMM_DEV` env should be provided when opening Metro bundler (and these commands do this unless already started or the `--no-bundler` flag is provided).

Test Plan:
- `yarn dev` still works the same way
- `expo run:ios` builds the native app successfully
- the same for `expo run:android`

Reviewers: ashoat, atul, jon

Reviewed By: ashoat, atul, jon

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D6309