[desktop] Add auto updating
Summary:
Uses the `autoUpdater` module and check for updates every 5 minutes. Exposes to the web app:
- current version
- listener for a new version
- method to trigger the update.

Depends on D6321

Test Plan: Tested with the later diffs that introduce UI in the web app. Created a few version of the app and checked if the download and automatic update is triggered correctly after uploading a new one (on both macos and windows).

Reviewers: inka, przemek, tomek

Reviewed By: tomek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6322