[electron update server] Path hazel to work with desktop-vX.Y.Z tags
Summary:
Because we have both mobile and desktop releases in our github repo we need to add a small change to the Hazel electron update server so that it only looks for releases that have a git tag that looks like `desktop-vX.Y.Z`.

Depends on D6353

Test Plan: Check if hazel correctly looks for `desktop-vX.Y.Z` git tags and ignores other tags.

Reviewers: atul, tomek, jon

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D6354