[native] Add presentation: 'transparentModal' to termsAndPrivacyModalScreenOptions
Summary: I should've covered this in D5891. Without this, I think the screen behind the modal won't appear on Android.

Test Plan: I haven't had a chance to test this, but my theory is that without this, the Terms & Privacy modal appears on Android with the background completely black. @kamil, have you seen this on Android?

Reviewers: kamil, tomek

Reviewed By: kamil

Subscribers: atul, kamil

Differential Revision: https://phab.comm.dev/D6330