[native] Make it possible to swipe back from anywhere
Summary:
The core change here is to set `gestureResponseDistance` and `swipeEdgeWidth` to a large value so they trigger anywhere on the screen.

After this change `SwipeableMessage` wasn't triggering anymore. In D304 I made it wait for the swipe-to-go-back gesture, but now that the swipe-to-go-back gesture is everywhere, the swipe-to-reply gesture didn't activate at all.

I removed the `waitFor` from D304 and it felt like things were okay. Perhaps something has changed in React Navigation since D304, but if not I think the fact that you swipe anywhere to go back now will make these gestures less likely to conflict.

After this diff, if we notice ourselves accidentally triggering swipe-to-reply when meaning to trigger swipe-to-go-back, we can revisit this decision.

Test Plan: Deployed dev and release builds to my iOS device and tried swiping around in a `MessageList` (triggering `Chat` gesture) and `ThreadList` (triggering `CommunityDrawer` gesture)

Reviewers: max, inka, tomek, bartek

Reviewed By: max

Subscribers: atul

Differential Revision: https://phab.comm.dev/D6331