[native] Avoid invariant crash in getDevNodeServerURL
Summary:
D6269 updated release builds to hit `getDevNodeServerURL`, which is causing a crash. This is happening because the result of `getDevNodeServerURL` is included in the default list of server options in the Developer tools section of the app.

There's probably no risk to calling this in production, especially since we already call `dev-hostname.js` code from release builds in `natNodeServer`. It's better to improve readability of the code and remove this `invariant` than to sidestep it.

Test Plan: Confirm release build no longer crashes

Reviewers: bartek

Reviewed By: bartek

Subscribers: tomek, atul

Differential Revision: https://phab.comm.dev/D6332