Move navigation panel from the sidebar to a topbar
Summary:
Linear issue: https://linear.app/comm/issue/ENG-2739/move-navigationpanel-from-the-sidebar-to-a-top-bar
Due to this change the `LeftLayoutAside` is no longer needed, and the styles for the `AppSwitcher`, `SettingsSwitcher` and `NavigationPanel` should be moved out of `left-layout-aside.css`. But this diff is quite big as it is, so I'll do that
in a following diff.

Test Plan:
Run web app, check that the top bar displayes correctly in `Inbox` and in `Settings`. Check that calendar filters display correctly, check that removing the calendar app works fine.

Reviewers:
przemek, michal, bartek

Subscribers: