[web] Remove LeftLayoutAside
Summary:
Since D6335 the `LeftLayoutAside` is obsolete. I'm removing it, and putting the `CommunityPicker` directly in the `App` component.
This also calls for removing the `letf-layout`aside.css`, which I'll do in the next diff, as it requires some code to be moved somewhere else.

Test Plan:
Run the web app, check that the sidebar displayes correctly.

Reviewers:
przemek, michal, bartek

Subscribers: