[web] Move topbar related code out of web/sidebar
Summary:
Since D6335 the `AppSwitcher`, `SettingsSwitcher` and `NavigationPanel` are no longer in a sidebar, but in a topbar. So the code related to them had to be moved outside of `web/sidebar`. Moreover since D6336 the `LeftLayoutAside` doesn't exist
anymore, so the `left-layout-aside.css` file had been renamed to `topbar.css`.

Test Plan: Run the web app, check that the topbar displayes correctly.

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6337