[web] Add flag to search bar, to avoid weird hot reload side effects
Summary:
Linear issue: https://linear.app/comm/issue/ENG-2775/white-frame-inside-of-search-bar
After some changes to a different part of the code, I was seeing a white frame inside of the search bar, due to a style from `styles.css` (for `textarea`) taking precedence over this style. This was happening most likely because of hot-reload.
It was happening when I was saving files or checking out a commit. But once I reloaded the page it was back to normal. It wasn't happening on production. So I'm adding this flag only so that other devs are not alarmed in the future.

Test Plan:
Run web app (dev mode) on a different commit. Checked out to this commit (this used to cause the frame to appear). Checked that the frame is not present.
Removing this `!important` flag (and saving the file, thus triggering a hot-reload) made the frame appear. Adding this flag back (and saving (...)) causes the frame to not appear. Checked this a couple of times.

Reviewers:
przemek, michal, bartek

Subscribers: