[web] Create a dummy topbar component
Summary:
issue: https://linear.app/comm/issue/ENG-2739/move-navigationpanel-from-the-sidebar-to-a-top-bar
Adding a `Topbar` component that will contain the `AppSwitcher` and the navigation info bar. It will be displayed in the `Inbox` tab.
For now it displays "dummy", because adding it's contents requires a lot of changes to the code. So in the interest of keeping the diff small it has a dummy content for now.

Test Plan:
Displayed it in the `App` component.

Reviewers:
przemek, michal, bartek
Subscribers: