[web] Add possibility for the navigation-panel to be horizontal
Summary:
issue: https://linear.app/comm/issue/ENG-2739/move-navigationpanel-from-the-sidebar-to-a-top-bar
When `AppSwitcher`(that uses `NavigationPanel` underneath) is used in the `Topbar` it'll need to be horizontal. The tabs will also have to display differently. But the `SettingSwitcher` is supposed to stay as it is.

Test Plan:
Pass `horizontal={true}` to `NavigationPanel.Container` in `AppSwitcher` and check that it displayes correctly.

Reviewers:
przemek, bartek, michal
Subscribers: