[web] Use SettingsSwitcher
Summary:
issue: https://linear.app/comm/issue/ENG-2739/move-navigationpanel-from-the-sidebar-to-a-top-bar
Because of how we use the grid, and because the designs call for the switcher in the settings tab to a sidebar, and in inbox tab to be a topbar, the `SettingsSwitcher` is now inside of the `AccountSettings` and `DangerZone`.
Our grid: https://github.com/CommE2E/comm/blob/master/web/style.css#L54
So the `SettingsSwitcher` has to be a part of `app` grid area.
Changing the way we use the grid - or not using the grid - can be done, but it seems like a really big side quest to my community drawer task. So for now I think this is a good enough solution.
On the other hand, it has been proposed this month on the monthly planning meeting, that we might want to make the web app more responsive to resizing. If we do decide take care of it, this could be improved in the process.

Test Plan: Run web app, check that the `SettingSwitcher` works correctly.

Reviewers: przemek, michal, bartek

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6346