[lib] Return style of text from getMessagePreview
Summary:
I noticed that `native/chat/message-preview.react.js` wasn't using `getMessagePreview`, even though much of the logic was the same.

I wanted to unify the codepaths, but there is was one thing that `native` needed that wasn't being returned by `getMessagePreview`: what style to render each of the parts as.

I also noticed that `web` was rendering styles slightly different that `native`, which will be fixed in the next diff. This diff doesn't actually use the newly-returned `style` field... it just updates the code to return it.

Test Plan: Tested in combo with following diffs

Reviewers: inka, ginsu, atul, tomek

Reviewed By: inka

Differential Revision: https://phab.comm.dev/D6357