[web] Use shades-black-70 for thread-preview-secondary
Summary:
In the last diff, I updated `web` behavior to match `native` and use this color when the last message on a read thread is robotext.

That change made the contrast on the robotext look bad. Previously we had this low contrast for just the username, but it seems more stark when it applies to a whole message.

I would rather use what's on `native`: `#999999` for `--thread-color-read` (primary), and `#555555` for `--thread-preview-secondary` (in dark mode). In contrast, on `web` today we're using `#808080` and `#404040` (respectively), the latter of which I update to `#666666` in this diff.

The downside is that the primary and secondary colors now seem a bit too similar. I wonder if we need an exception to the design system here?

Depends on D6358

Test Plan:
| before | after |
| {F340663} | {F340664} |
(identical except "ashoat added commbot" near the bottom, and usernames prefixed to message previews)

Reviewers: inka, ginsu, atul, tomek

Reviewed By: atul

Differential Revision: https://phab.comm.dev/D6359