[native] Use getMessagePreview in native MessagePreview
Summary:
After D6357 updated `getMessagePreview` to return a style for each portion, we can update `native` to use that function to unify codepaths with `web`.

This should be a no-op. Compare to `web`'s version of `MessagePreview` in D6358 – very similar code.

Depends on D6359

Test Plan:
| before | after |
| {F340667} | {F340668} |
(should be identical)

Reviewers: inka, ginsu, atul, tomek

Reviewed By: atul

Differential Revision: https://phab.comm.dev/D6360