[web] Move code around to match the new code logic
Summary:
issue: https://linear.app/comm/issue/ENG-2739/move-navigationpanel-from-the-sidebar-to-a-top-bar
Now that the `SettingSwitcher` is no longer in the sidebar, it's code had to be moved outside of `web/sidebar`. The styles in `left-layout-aside.css` were only used for `SettingsSwticher` (and
horizontal `NavigationPanel` that is used only in `SettingsSwitcher` so the file could be renamed and also moved. `NavigationPanel` is used for `SettingSwitcher` and `AppSwitcher`, so only for
navigation panels. Thus all of this code could be put into `web/navigation-panels` as well as the `Topbar`, that is a navigation panel (contains `AppSwitcher` and will contain the navigation state
information).

Test Plan:
Run web app, check that the top bar in `Inbox`, and the `SettingSwitcher` in `Settings` display correctly. Run `flow check` in `web/`.

Reviewers:
przemek, michal, bartek
Subscribers: