[web] Make drawer items not expandable in the calendar tab
Summary:
issue: https://linear.app/comm/issue/ENG-2789/make-drawer-items-in-calendar-tab-not-expandable
The new designs specify, that drawer items are not supposed to be expandable in the Calendar tab.

Test Plan:
Run the web app, check that drawer items are not expandable in the Calendar tab, but stay expandable in outher tabs.

Reviewers:
tomek, przemek, michal, bartek

Subscribers: