Implement native method to get initial notification and use it in JavaScript
Summary:
This differential implements native method to get initial notification - notification which delivery and tapping booted the app - and uses it in JavaScript. CommAndroidNotifications module
typing is updated as well.

Test Plan: Add logging to log initialNotif from JavaScript. Build the app. Kill it. Send notification. Tap it. Ensure the content of notification is displayed in metro console and it is valid.

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6366