[keyserver] Don't crash on malformed socket input
Summary:
This diff makes it so now if `checkInputValidator` throws a `ServerError` due to input validation failing, `clientSocketMessage` will be set in this `catch` block.

This is important because the `catch` block has an `invariant` expecting the `clientSocketMessage` to be set.

Test Plan: I had a reproducible issue where I could trigger `checkInputValidator` to throw a `ServerError`. I confirmed that the `invariant` was no longer triggered after my change

Reviewers: tomek, atul

Reviewed By: atul

Differential Revision: https://phab.comm.dev/D6370