Implement native method to check for notifications permissions and use it in JavaScript
Summary: This differential implements native method to check for notifications permissions and uses it in JavaScript. Relevant types are updated as well.

Test Plan:
Build the app and open it. Open settings and revoke comm notifications permissions. Go back to app and ensure that "Failed to register push notifs" message is displayed and no notifications
are delivered. Close the app. Go back to settings, bring notifications permissions back. Open the app, ensure that previous message is not displayed and notifications are succesfully delivered.

Reviewers: tomek, bartek

Reviewed By: tomek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6380