Implement native method to get device token and use it in JavaScript
Summary: This differential implements native method to get notifications token and uses it in JavaScript. Relevant types are updated as well

Test Plan:
Build app but make sure you are not installing fresh. Notifuications token is send from CommNotificationsHandler when app is installed, so it would interfere with this test plan. Make sure all
 notifications functionality works. Without ability to get notifications token it won't.

Reviewers: tomek, bartek

Reviewed By: tomek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6381