[web] Highlight drawer items
Summary:
issue: https://linear.app/comm/issue/ENG-2738/highlight-drawer-item
Drawer items are supposed to be highlighted in the Chat tab, and not highlighted in the Calendar tab. The Apps tab and user settings tabs for now behave like the Chat tab, because it is still being
discussed what we want to do there (https://linear.app/comm/issue/ENG-2789/make-drawer-items-in-calendar-tab-not-expandable)

Test Plan:
Run the web app, check that drawer items are highlighted in the Chat, and are not in the Calendar App.
Reviewers:

Subscribers: