[web] switch like button to react button on message tooltip
Summary:
Introduced the emoji keyboard and make it render whenever `renderEmojiKeyboard` in the tooltip context is set to true. Also modified `useMessageReactAction` to return a react button instead of a like button. Will be working on improving the positioning of the emoji keyboard in a subsequent diff

---
Depends on D6401
Linear Task: [[ https://linear.app/comm/issue/ENG-2778/replace-like-button-in-message-tooltip-with-a-react-button | ENG-2778 ]]

Test Plan:
Please watch the video below to see a demo of how the new react button works

{F346108}

Reviewers: atul, tomek

Reviewed By: atul

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6402