[web] Add `@matrix-org/olm` dependency
Summary:
We need the `Olm` dependency in order to get device public key to include in SIWE message. (We may end up using the dependency for more things later on, but this is the only use case atm)

Including this dependency in browser isn't documented well, so I tried to get it to work based off of what I saw in other repos (eg jitsi-meet, element-web)

Test Plan:
Was able to import the dependency in `web/account/siwe.react.js` and run `window.Olm.init()` without issue.

Tested with `dev` and `prod` to ensure that we're able to `init()` Olm as expected.

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D6400