[keyserver/native] Attach a threadID to each media (image/video) upload on native
Summary: For the native side of things, we need to populate the thread column when creating a new entry in the table. I pass in the thread id as part of extras (type MultimediaUploadExtras).

Test Plan: Upload several media (images and videos) and check the `uploads` table to confirm that the `thread` column is populated.

Reviewers: atul, ginsu, przemek, tomek

Subscribers: