[web] introduce renderEmojiKeyboard and setRenderEmojiKeyboard to tooltip-provider
Summary:
introduce `renderEmojiKeyboard` and `setRenderEmojiKeyboard` to `tooltip-provider`. `renderEmojiKeyboard` and `setRenderEmojiKeyboard` will be used to condionally render the emoji keyboard whenever the user clicks on the react button on the message tooltip. I will be introducing the react button in a subseqent diff.

---

Depends on D6398
Linear Task: [[ https://linear.app/comm/issue/ENG-2776/implement-web-emoji-keyboard-library | ENG-2776 ]]

Test Plan: `flow` and this diff will be further tested in a subsequent diff

Reviewers: atul, tomek

Reviewed By: atul

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6399