[lib] Show cached ENS names on first render
Summary:
Instead of waiting for `ENSCache` to resolve a `Promise`, this diff makes it so if an ENS name is cached, we show it immediately from `useStringForUser` on the first render.

Depends on D6418

Test Plan: I added a `console.log` of the result of `useStringForUser`, and then logged out and back in on the iOS simulator. After this diff, the first render of `useStringForUser` was correct when the result was cached, but before this diff the first render had the ETH address instead of the ENS name

Reviewers: atul

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D6419