[web] Include `primaryIdentityPublicKey` in SIWE message
Summary:
Include `primaryIdentityPublicKey` from `olm` in the SIWE message as we do on `native`. Note that this diff just includes getting the `primaryIdentityPublicKey` from `olm` and including it to the SIWE message statement.

NOTE: Persistence of the `primaryIdentityPublicKey` will happen in a subsequent diff.

Test Plan:
Works as expected, can see that the `primaryIdentityPublicKey` is included in the SIWE message and everything is validated on the `keyserver` side!

{F346483}

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D6410