[native] Convert RelationshipList to hook
Summary: We need to call `setENSNames` on some things that are derived from the component's state. To access the state from a hook, we need to convert the whole component to a hook.

Test Plan: Tested and played around with `RelationshipList` in iOS Simulator. Added friend, searched for friend, removed friend. Also Flow

Reviewers: atul

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D6435