[native] Convert ComposeSubchannel to hook
Summary: We need to call `setENSNames` on some things that are derived from the component's state. To access the state from a hook, we need to convert the whole component to a hook.

Test Plan: Tested and played around with `ComposeSubchannel` in iOS Simulator. Selected an ENS user, selecting a non-ENS user, created a subchannel. Verifed button enabled/disabled as expected. Also Flow

Reviewers: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D6439