[native] introduce reaction selection popover component
Summary:
introduce the `ReactionSelectionPopover` component. For the default emojis, I could not fit all the emojis menioned in the linear task in the popover. So I went with FB messengers default emojis for
now, but we can always change it in the future

Test Plan: Please watch the demo video to see what the reaction selection popover component looks like

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6454