[web] Add to calendar filters reducer logic for the COMMUNITY filters
Summary:
issue: https://linear.app/comm/issue/ENG-2794/make-it-possible-to-display-only-chats-from-a-given-community-in-the
In D6457 I added architectural changes that allow for a new type of calendar filter. In this diff I address these changes in the `CalendarFiltersReducer`, by handling the new type of action. What we want to
happen, is for the calnderFilters in the redux store, to be set to: the calendra filters of other types as they used to be + calendar filter of type COMMUNITY to have threadIDs equal to threadIDs
from the actions payload.

Test Plan:
Dispatch `updateCalendarCommunityFilter` and `clearCalendarCommunityFilter` actions, check in redux tools that they succeed, and redux state changes as expected.
Reviewers:
kamil, tomek

Subscribers: